+++ docs/source/conf.py
@@ -15,6 +15,8 @@
 import sys
 import os
 import re
+import logging
+import types
 # to let sphinx find the actual source...
 sys.path.insert(0, os.path.abspath("../.."))
 from webhook2lambda2sqs.version import VERSION
@@ -291,7 +293,7 @@ texinfo_documents = [
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
-    'https://docs.python.org/2/': None,
+    'https://docs.python.org/2.7/': None,
     'boto3': ('http://boto3.readthedocs.io/en/latest', None),
     'botocore': ('http://botocore.readthedocs.io/en/latest', None)
 }
@@ -312,17 +314,37 @@ linkcheck_ignore = [
     r'https?://docs\.aws\.amazon\.com/apigateway/.*'
 ]
 
+
 # exclude module docstrings - see http://stackoverflow.com/a/18031024/211734
 def remove_module_docstring(app, what, name, obj, options, lines):
     if what == "module":
         del lines[:]
 
+
 # ignore non-local image warnings
 def _warn_node(self, msg, node, **kwargs):
-    if not msg.startswith('nonlocal image URI found:'):
-        self._warnfunc(msg, '%s:%s' % get_source_line(node))
+    if msg.startswith('nonlocal image URI found:'):
+        return
+    self._warnfunc(msg, '%s:%s' % get_source_line(node))
 
 sphinx.environment.BuildEnvironment.warn_node = _warn_node
+
+
+# BEGIN workaround for https://github.com/sphinx-doc/sphinx/issues/3860
+def _images_log_warning(self, msg, *args, **kwargs):
+    if msg.startswith(
+            'Could not fetch remote image: '
+            'https://readthedocs.org/projects/webhook2lambda2sqs/badge/?'
+    ):
+        print('Suppressing RTD badge remote image warning: %s' % msg)
+        return
+    self._orig_warning(msg, *args, **kwargs)
+
+img_log = logging.getLogger('sphinx.transforms.post_transforms.images')
+img_log._orig_warning = img_log.warning
+img_log.warning = types.MethodType(_images_log_warning, img_log)
+# END workaround for https://github.com/sphinx-doc/sphinx/issues/3860
+
 
 def setup(app):
     app.connect("autodoc-process-docstring", remove_module_docstring)