+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.2.0 (2017-06-25)
+------------------
+
+* Fix terraform output handling for 0.8.x
+* Add Python 3.6 to test environment list
+* Make compatible with Terraform 0.9+
+* Add hack to ``docs/source/conf.py`` as workaround for https://github.com/sphinx-doc/sphinx/issues/3860
+
 0.1.0 (2016-08-06)
 ------------------